Split verify command into sub directories
To accommodate following addition of the Voyager block explorer
verification.
